エネミー関連のリファクタリングはほとんど終わったと思われ

+++ Shooters!/Enemy.cpp
@@ -2,7 +2,7 @@
 #include"DxLib.h"
 #include"Random.h"
 
-Enemy::Enemy(Vec2 &pos, int graph_handle, int hp, int speed, int range, int anime_interval, int shot_speed)
+Enemy::Enemy(Vec2 &pos, int graph_handle, int hp, float speed, int range, int anime_interval, int shot_speed)
 	: range(range), anime_interval(anime_interval), shot_speed(shot_speed),
 	draw_flag(true), erase_flag(false), pos(pos), hp(hp), speed(speed), counter(0),
 	refresh_time(GetNowCount()), refresh_time_a(0), anime_switch_flag(true), graph_index(0),
